requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
cloudinary==1.36.0
Werkzeug==3.0.1
gunicorn==21.2.0